NONE: Refactor consistency mocks
During one of the PR's, we noticed consistency mocks are becoming too complex to use.
For example, there's no way to exclude supplier call for PermanentConsistency. For
PermanentInconsistency it's possible, but the parameter is passed via the constructor,
and it's not easy to read. This change is an attempt to make it simpler, more consistent,
and to give more flexibility.