Avoid unnecessary switch to `MainConnection` while setting `read-only` mode
As per JavaDoc `read-only` mode is just a hint to the driver to enable
optimisations. In the case of `DualConnection` switching to the main just in case
is anti-optimisation.

I found out we set it to `false` just in case in `DatabaseAccessorImpl#runInTransaction`.
There might be more places like this, that prevent us from utilising replica more.