Remove circuit breaker
There's no API for circuit breaker, so we should not enforce `BreakOnNotSupportedOperations`.
No circuit breaker is better because it allows users to implement circuit breaker
on an application level.

We also have feedback from other Atlassian teams that `BreakOnNotSupportedOperations`
doesn't fit their needs and they can't override it.

The other option would be to open CircuitBreaker API, but in past,
we decided not to publish CircuitBreaker in the current form as we started to work
on a better SPI (https://github.com/atlassian-labs/db-replica/pull/11). Better SPI is on hold now.