Allow to opt-out from replication
Scenarios when we want to opt-out from the replica:
- we have special contexts where we want to load data from
the main database only (to be 100% sure the data is always consistent, regardless of ReplicaConsistency implementations).
- limit the load that goes to the replica.
- turn off calls to a replica without waiting for DualConnections to be closed.

Until now, we were using `ReplicaConsistency#isConsistent` to guard
calls to a replica. It works, but it also is confusing (why replica is not consistent in this context?). The goal of
this change is to provide a dedicated SPI.