JCES-2252: Increment LSN when new value appears
When LSN fetches slowed down and didn't fetch new LSN within a timeout,
the result of the fetch was discarded and all the requests were waiting
for the new value to appear. Only an SQL query that meets the timeout
criteria could update the value. The LSN timing out was also clearing
the lock, so the successful queries were unable to set the new value.
Not increasing the cache value could lead to more requests checking the
LSN and to create a vicious circle affecting customers for a longer period
of time.